+++ hrp/cmd/pytest.go
@@ -1,17 +1,17 @@
 package cmd
 
 import (
-	"fmt"
 	"strings"
 	"time"
 
+	"github.com/httprunner/funplugin/myexec"
+	"github.com/pkg/errors"
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
 
-	"github.com/httprunner/httprunner/v4/hrp/internal/myexec"
+	"github.com/httprunner/httprunner/v4/hrp/internal/code"
 	"github.com/httprunner/httprunner/v4/hrp/internal/pytest"
 	"github.com/httprunner/httprunner/v4/hrp/internal/sdk"
-	"github.com/httprunner/httprunner/v4/hrp/internal/version"
 )
 
 var pytestCmd = &cobra.Command{
@@ -29,13 +29,11 @@ var pytestCmd = &cobra.Command{
 			})
 		}()
 
-		packages := []string{
-			fmt.Sprintf("httprunner==%s", version.HttpRunnerMinimumVersion),
-		}
+		packages := []string{"httprunner"}
 		_, err = myexec.EnsurePython3Venv(venv, packages...)
 		if err != nil {
 			log.Error().Err(err).Msg("python3 venv is not ready")
-			return err
+			return errors.Wrap(code.InvalidPython3Venv, err.Error())
 		}
 		return pytest.RunPytest(args)
 	},