+++ hrp/internal/version/init.go
@@ -6,6 +6,3 @@ import (
 
 //go:embed VERSION
 var VERSION string
-
-// httprunner python version
-const HttpRunnerMinimumVersion = "v4.3.5"